+++ radiomesh/tests/test_wgrid.py
@@ -41,6 +41,6 @@ def test_numba_wgrid(nx, ny, fov, oversampling):
   vis_grid, weight_grid = wgrid(uvw, vis, weights, flags, freqs, Datum(wgrid_params))
 
   assert vis_grid.shape == (4, nx, ny)
-  assert weight_grid.shape == (nx, ny)
-  assert vis_grid.dtype == vis.real.dtype
+  assert weight_grid.shape == (4, nx, ny)
+  assert vis_grid.dtype == vis.dtype
   assert weight_grid.dtype == weights.dtype